% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_Modified_probabilities}
\alias{compute_Modified_probabilities}
\title{Compute Modified Item Response Probabilities}
\usage{
compute_Modified_probabilities(results, theta_all)
}
\arguments{
\item{results}{The data frame returned by fit_binary_irt.}

\item{theta_all}{A numeric vector representing ability levels (total scores).}
}
\value{
A matrix of predicted probabilities.
}
\description{
Compute Modified Item Response Probabilities
}
