% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_rasch_difficulties_ordered}
\alias{extract_rasch_difficulties_ordered}
\title{Extract Rasch Item Difficulties in Original Order}
\usage{
extract_rasch_difficulties_ordered(final_logit_matrix)
}
\arguments{
\item{final_logit_matrix}{The matrix returned by rasch_logit.}
}
\value{
A data frame of item difficulties.
}
\description{
Extract Rasch Item Difficulties in Original Order
}
