% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_binary_irt}
\alias{fit_binary_irt}
\title{Fit Binary IRT Model using GLM}
\usage{
fit_binary_irt(data_mat, total_score)
}
\arguments{
\item{data_mat}{A numeric matrix of responses.}

\item{total_score}{A numeric vector of total scores.}
}
\value{
A data frame containing Intercept, Slope, and Threshold for each item.
}
\description{
Fit Binary IRT Model using GLM
}
