% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_item_curves}
\alias{plot_item_curves}
\title{Plot Item Characteristic Curves (ICC)}
\usage{
plot_item_curves(theta_all, Modified_prob_matrix, results)
}
\arguments{
\item{theta_all}{A numeric vector of ability levels.}

\item{Modified_prob_matrix}{The matrix returned by compute_Modified_probabilities.}

\item{results}{The data frame returned by fit_binary_irt.}
}
\description{
Plot Item Characteristic Curves (ICC)
}
