% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_rasch_curves}
\alias{plot_rasch_curves}
\title{Plot Rasch Item Curves}
\usage{
plot_rasch_curves(prob_matrix, final_logit_matrix)
}
\arguments{
\item{prob_matrix}{The matrix returned by compute_Modified_probabilities.}

\item{final_logit_matrix}{The matrix returned by rasch_logit.}
}
\description{
Plot Rasch Item Curves
}
