% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rasch_logit}
\alias{rasch_logit}
\title{Compute Logit and Row Means}
\usage{
rasch_logit(prob_matrix)
}
\arguments{
\item{prob_matrix}{The matrix returned by compute_Modified_probabilities.}
}
\value{
A matrix of logits with an added column for row means (Student Ability).
}
\description{
Compute Logit and Row Means
}
